# Avoid race condition