There seems to be some race-condition in `build.rs`
that causes random build results to be produced for `/usr/bin/difft`.

This change seems to have negligible influence on overall build time.

If you make a better patch, I can test it.