@bmwiedemann 2025-02-10
It produced different binary results. Probably all working, but did not test. I think, some elements were ordered differently.

You can use https://github.com/bmwiedemann/reproducibleopensuse/blob/master/filterdiff with 'strings' and difft binaries as params. (and the diff is probably more readable with `lto = "off"` )

I measured build times on a 2x16 core 2.6GHz host with build VMs:
without patch on 1-core-VM: 657s
with patch on 1-core-VM: 654s
without patch on 4-core-VM: 142s
with patch on 4-core-VM: 175s

So the patch does slightly speed up single-core builds but slow down multi-core builds. If we can avoid the variations in other ways, that would be good.

`-Wno-unused-parameter` seems to not make a difference.