@Wilfred 2025-05-07
Could you try setting `RAYON_NUM_THREADS` to 1? That might be sufficient to fix this.