@bmwiedemann 2025-05-08
Indeed, this seems to produce the same binaries as with this PR applied.